fixup! fixup! tmp

+++ nix/mobile/android.nix
@@ -107,6 +107,7 @@ let
       cp -a ${nodePackage}/lib/node_modules/${nodeProjectName}/node_modules .
       chmod u+w -R ./node_modules/react-native
 
+      # Set up symlinks to mobile enviroment in project root 
       for f in package.json .babelrc VERSION metro.config.js
       do
         rm -f $f && ln -s ./mobile_files/$f
@@ -114,13 +115,14 @@ let
     '';
     patchPhase = ''
       # Patch maven and google central repositories with our own local directories. This prevents the builder from downloading Maven artifacts
-      ${lib.concatStrings (lib.mapAttrsToList (projectName: deriv: ''
-      substituteInPlace ${lib.optionalString (projectName != nodeProjectName) "node_modules/${projectName}/"}android/build.gradle \
-        --replace "google()" "maven { url \"${deriv}\" }" \
-        --replace "jcenter()" "maven { url \"${deriv}\" }" \
-        --replace "https://maven.google.com" "${deriv}" \
-        --replace "\$rootDir/../node_modules/react-native/android" "${mavenLocalRepos.react-native-android}"
-      '') (lib.filterAttrs (name: value: name != "react-native-android") mavenLocalRepos))}
+      ${lib.concatStrings (lib.mapAttrsToList (projectName: deriv:
+        let targetGradleFile = "${lib.optionalString (projectName != nodeProjectName) "node_modules/${projectName}/"}android/build.gradle";
+        in ''
+      grep 'google()' ${targetGradleFile} > /dev/null && substituteInPlace ${targetGradleFile} --replace "google()" "maven { url \"${deriv}\" }"
+      grep 'jcenter()' ${targetGradleFile} > /dev/null && substituteInPlace ${targetGradleFile} --replace "jcenter()" "maven { url \"${deriv}\" }"
+      grep 'https://maven.google.com' ${targetGradleFile} > /dev/null && substituteInPlace ${targetGradleFile} --replace "https://maven.google.com" "${deriv}"
+      grep '\$rootDir/../node_modules/react-native/android' ${targetGradleFile} > /dev/null && substituteInPlace ${targetGradleFile} --replace "\$rootDir/../node_modules/react-native/android" "${mavenLocalRepos.react-native-android}"
+        '') (lib.filterAttrs (name: value: name != "react-native-android") mavenLocalRepos))}
 
       # Patch prepareJSC so that it doesn't try to download from registry
       substituteInPlace node_modules/react-native/ReactAndroid/build.gradle \
@@ -131,11 +133,8 @@ let
       substituteInPlace scripts/build_no.sh \
         --replace "(git rev-parse --show-toplevel)" "STATUS_REACT_HOME"
 
-      # TODO: figure out why we get `path may not be null or empty string. path='null'`
-      substituteInPlace node_modules/react-native/ReactAndroid/release.gradle \
-        --replace "classpath += files(project.getConfigurations().getByName(\"compile\").asList())" ""
-
       # HACK: Run what would get executed in the `prepare` script (though index.js.flow will be missing)
+      # Ideally we'd invoke `npm run prepare` instead, but that requires quite a few additional dependencies
       (cd ./node_modules/react-native-firebase && \
        chmod u+w -R . && \
        mkdir ./dist && \
@@ -152,6 +151,16 @@ let
         LD_LIBRARY_PATH=$LD_LIBRARY_PATH:${stdenv.lib.makeLibraryPath [ zlib ]} \
           gradle --no-daemon react-native-android:installArchives
       )
+
+      # If these directories are missing, there will be an error later on when building with this node_modules directory:
+      # What went wrong:
+      # Failed to create parent directory 'node_modules/react-native-touch-id/android/build' when creating directory '/home/pedro/src/github.com/status-im/status-react/node_modules/react-native-touch-id/android/build/intermediates/check_manifest_result/release/checkReleaseManifest/out'
+      for p in react-native-{background-timer,camera,config,dialogs,firebase,fs,http-bridge,image-resizer,image-crop-picker,keychain,languages,mail,securerandom,shake,splash-screen,status-keycard,svg,touch-id,webview,webview-bridge} \
+               realm
+      do
+        chmod -R u+w node_modules/$p
+        mkdir -p node_modules/$p/android/build
+      done
     '';
     installPhase = ''
       rm -rf $out
@@ -163,10 +172,14 @@ let
         --replace "packageReactNdkLibs(dependsOn: buildReactNdkLib, " "packageReactNdkLibs(" \
         --replace "../../../deps/${jsc-filename}.tar.gz" "${react-native-deps}/deps/${jsc-filename}.tar.gz" 
     '';
-    dontPatchELF = true; # The ELF types are incompatible with the host platform, so let's not even try
+
+    # The ELF types are incompatible with the host platform, so let's not even try
+    dontPatchELF = true;
     noAuditTmpdir = true;
+
+    # Take whole sources into consideration when calculating sha
+    outputHashMode = "recursive";
     outputHashAlgo = "sha256";
-    outputHashMode = "recursive"; # Take whole sources into consideration when calculating sha
   };
 
 in
@@ -177,61 +190,50 @@ in
     shellHook =
       androidEnvShellHook + ''
       $STATUS_REACT_HOME/scripts/generate-keystore.sh
-
+    '' +
+    # Check if we need to copy node_modules (e.g. in case it has been modified after last copy)
+    ''
+      needCopyModules=1
       if [ -d ./node_modules ]; then
-        chmod u+w -R ./node_modules
-        rm -rf ./node_modules || exit
+        if [ -f ./node_modules/.copied ]; then
+          echo "Checking for modifications in node_modules..."
+          modifiedFiles=$(find ./node_modules -writable -type f -newer ./node_modules/.copied)
+          if [ $? -eq 0 ] && [ -z "$modifiedFiles" ]; then
+            needCopyModules=0
+            echo "No modifications detected."
+          fi
+        fi
+        if [ $needCopyModules -eq 1 ]; then
+          chmod u+w -R ./node_modules
+          rm -rf ./node_modules || exit
+        fi
       fi
-      echo "Copying node_modules from Nix store (${deps}/node_modules)..."
-      # mkdir -p node_modules # node_modules/react-native/ReactAndroid
-      time cp -HR --preserve=all ${deps}/node_modules .
-      echo "Done"
-
+      if [ ! -d ./node_modules ]; then
+        echo "Copying node_modules from Nix store (${deps}/node_modules)..."
+        time cp -HR --preserve=all ${deps}/node_modules . && \
+          chmod u+w ./node_modules && \
+          touch ./node_modules/.copied && \
+          chmod u-w ./node_modules
+        echo "Done"
+      fi
+    '' +
+    # Fix permissions in certain directories so that React Native doesn't have a fit
+    ''
       rndir='node_modules/react-native'
       rnabuild="$rndir/ReactAndroid/build"
+      chmod u+w -R $rnabuild
       chmod 744 $rndir/scripts/.packager.env \
                 $rndir/ReactAndroid/build.gradle \
                 $rnabuild/outputs/logs/manifest-merger-release-report.txt \
                 $rnabuild/intermediates/library_manifest/release/AndroidManifest.xml \
-                $rnabuild/intermediates/aapt_friendly_merged_manifests/release/processReleaseManifest/aapt/AndroidManifest.xml \
-                $rnabuild/intermediates/aapt_friendly_merged_manifests/release/processReleaseManifest/aapt/output.json \
-                $rnabuild/intermediates/incremental/packageReleaseResources/compile-file-map.properties \
-                $rnabuild/intermediates/incremental/packageReleaseResources/merger.xml \
+                $rnabuild/intermediates/aapt_friendly_merged_manifests/release/processReleaseManifest/aapt/{AndroidManifest.xml,output.json} \
+                $rnabuild/intermediates/incremental/packageReleaseResources/{merger.xml,compile-file-map.properties} \
                 $rnabuild/intermediates/merged_manifests/release/output.json \
                 $rnabuild/intermediates/symbols/release/R.txt \
                 $rnabuild/intermediates/res/symbol-table-with-package/release/package-aware-r.txt
-      chmod u+w -R $rnabuild
-
-      # What went wrong:
-      # Failed to create parent directory '/home/pedro/src/github.com/status-im/status-react/node_modules/react-native-touch-id/android/build' when creating directory '/home/pedro/src/github.com/status-im/status-react/node_modules/react-native-touch-id/android/build/intermediates/check_manifest_result/release/checkReleaseManifest/out'
-      chmod u+w node_modules
-      for p in react-native-background-timer \
-               react-native-camera \
-               react-native-config \
-               react-native-dialogs \
-               react-native-firebase \
-               react-native-fs \
-               react-native-http-bridge \
-               react-native-image-resizer \
-               react-native-image-crop-picker \
-               react-native-keychain \
-               react-native-languages \
-               react-native-mail \
-               react-native-securerandom \
-               react-native-shake \
-               react-native-splash-screen \
-               react-native-status-keycard \
-               react-native-svg \
-               react-native-touch-id \
-               react-native-webview \
-               react-native-webview-bridge \
-               realm
-      do
-        chmod -R u+w node_modules/$p
-        mkdir -p node_modules/$p/android/build
-      done
-      chmod u-w node_modules
-
+    '' +
+    # Lastly, add some configuration
+    ''
       export PATH="$PATH:${deps}/node_modules/.bin"
       export GRADLE_USER_HOME="$STATUS_REACT_HOME/.gradle" # Assign our own Gradle cache location, so that other non-Nix projects don't pollute our environment
     '';
